Ralentissement vitesse rotation chien

+++ Project_SDL_Part1_base/Project_SDL1.h
@@ -105,7 +105,7 @@ public:
   SDL_Surface* window_surface_ptr_; // ptr to the surface on which we want the
   shepherd_dog(SDL_Surface* window_surface_ptr, const shepherd& master);
   ~shepherd_dog();
-  void move(const shepherd& master, int degree);
+  void move(const shepherd& master, double degree);
 
 private:
   SDL_Surface* image_ptr_; // The texture of the sheep (the loaded image), use
